Simulation study using grid search

Performs a simulation study to assess the performance of the test. Depending on user inputs this could take a while since Rho is estimated using a grid search from -.99:.99:.01

simulation_study_grid_ar1(
  outDir = here::here("out.txt"),
  betaVec = c(0.12, 0.25, 0.5),
  rhoVec = c(0, 0.25, 0.5, 0.75, 0.95),
  sigmaVec = c(0.25, 0.5, 0.75),
  nTVec = c(10),
  nSims = 200,
  nBootSims = 500,
  rhoSearch = NULL
)

Arguments

outDir

Character string. Path to output file

betaVec

Numeric vector. Values for beta_1 (slope/trend parameter)

rhoVec

Numeric vector. Values for autoregressive parameter

sigmaVec

Numeric vector. Values of standard deviation of noise

nSims

Numeric scalar. Number of time series to simulate

nBootSims

Numeric scalar. Number of bootstrap data sets

rhoSearch

Numeric Vector. Values to use to for rho in grid search eg (seq(-.99,.99,.01))

nVec

Numeric vector. Values for the length of time series to simulate

Details

For convenience the intercept, beta_0 is set to zero

Examples

if (FALSE) {
simulation_study_grid(here::here("output.txt"),betaVec = 0.5,rhovec =0.5, sigmaVec = 0.25,
nTVec = 10, nSims = 200, nBootSims = 500)
}